refactor: extract function to retrieve data from database
